Port shows index page to Jekyll
Now per-year show separation is organized not into standalone pages, but
into a Jekyll collection, so we can generate the index page taking all
card attributes from the corresponding collection entries.

Also each thumbnail on a card has an alternate text. For now it is
preset statically by URL, but later it will be generated by show details
like a caption in a place where the photo is originally taken from.

For a smooth transition to a long process of per-year show pages porting,
a content of the pages was taken with no modification. It can not be
even wrapped into a default template because of a difference between an
old stylesheet and a new one.