Replace more photos with their originals
Extra original photos straight from a camera with proper shooting date
and time are found.

One of such photos (img/2008/06291.jpg) is equivalent to an unreferenced
dead weight (img/photos/hillari/37.jpg) that existed from the very
first commit of a repository and for some reason was omitted from
an incremental list in 20a4971:females/hillari/photos.htm.

Another (img/2009/07031) has two variants nearby in the same gallery,
one is like an original and another is cropped. The cropped version is
removed here to be regenerated in the future if necessary.