Migrate images into a resource directory
Initially, here was an idea to use symlinks to ease migration from a
root-based directory for Jekyll to a res-based directory for Python
scripts. The end of the transition was planned after all legacy
previews are replaced with proper photos. However, GitHub Actions does
not allow symbolic links because of its container configuration, so
the transition became forced.