Make test and deploy parallel
Initially, there was an idea to not allow to upload a version of a site
that has some broken links or does not satisfy some metrics. However,
due to a large artifact (~2000 files), such an approach makes build
slow.

Considering a non-fatal nature of non-template broken links or failed
metrics we can deploy a site even with failed metrics, so this commit
drastically speeds up build time from ~10m to something more manageable.