Delete legacy previews
Initially, a preview was preserved if no original photo is available
to maintain quality. However, a necessity to use the same single-size
preview in multiple cases that require various widths or heights defeats
this idea.

So this commit deletes all "*-p.jpg" previews along with their special
handling in a script.