Fix show header display
An initial idea was to write date for each event inside the same day
to increase quality of search indexation of these events. However,
it made event headers unreadable. As an attempt to increase readability,
remove an extra "г." city prefix that is a repetition of a year suffix
"г." and disconnect photos from a header.