Add simplified syntax for no-photo pedigree cells
Before porting from YAML to CSV it was easy to leave a pedigree cell
without a photo parameter and forget about it because of loose
and vertically stretched format, and impossibility to grep records with
missing "photo" parameter. So a special value "none" was required
to grep it further.

After porting to compact, tabular CSV we got the opposite situation:
an ability to easily spot empty cells but clutter with special values.

So this commit replaces a special value for an absent photo with just
an empty cell.