Fix interpretation of dog types
Initially a type field was though as an usual dog attribute.

However, it is better to treat type as a special mark that moves a dog
from a pile of other dogs into some special category. The rest of a pile
is sorted out by heuristics. For example, a non-marked dog is displayed
only if it has a date of birth, otherwise it is treated as a pseudo
record with minimal data for pedigrees.

Before there was a plan to get rid of the type column at all and define
breeders as dogs with children. However, in reality breeders are dogs
that we want to advertise for breeding; they can have no children yet
or have untracked children in other kennels.