Migrate photos outside auto galleries
Here is a plan to generate photo titles from dates encoded in file names
and actors enumerated somewhere in an image directory. So start
migration of photos which only mention will not be wiped off with
res/dogphotos.yml along with parallel migration of every photo to date
naming scheme.