Move a misyeared photo out of search path
For 2021 shows, a cropped photo from 2020 (2020/09201 d=tekila.jpg) is
used because of lack of better 2021 photo yet. To avoid its insertion
into a gallery, move it outside of path scanned for photos.

Since we moved a photo outside of a year directory, a date extraction
for a gallery fails, so extend fallback coverage from "file is not added
via shows so not known yet" to "something is wrong with processing".