Move photos so photos/*.jpg will not match */*.*
To automatically extract photos and their attributes from img/
we are going to use globbing with "*/*.*" pattern, where the first
asterisk is for a year, the second is for a month, a day, a sequence
number and attributes. However, photos right inside photos/ folder are
also captured even if they have neither date nor attributes.

Previously it was explicitly filtered out, but now we just need to move
generic dog photos into a special directory for such ones.