Make artifact list generators return pathlib.Path
This commit starts porting patches from strings to Path. It allows to
eliminate errors when arbitrary strings are passed to places where path
strings are expected.

To isolate changes made in the very beginning of generation conveyor,
the following step in build_site.py stringifies the object to pass it
to not-yet-updated modules.