Make generation of navigation benefit from pathlib
The only place tools.Document uses a path is to determine what item of
a navigation bar corresponds to the current document.

By the way, introduction of Path made one more important simplification.
Since "index.html" is omitted from URLs by convenience, a code that
stripped it away used regex substitution. However,

> Some people, when confronted with a problem, think "I know, I'll use
> regular expressions." Now they have two problems.
>
> -  Jamie Zawinski

So replacement of such a powerful but easily coming of control method
along with the only place where re module is used, made a script easier
for understanding.