Decouple photo creation and attributes assignment
Currently, photo objects are created with all attributes ready. However,
the next commit will introduce mapping of attributes to lists of photo
objects. This mapping is established while attribute scanning is in
progress, so there is nothing to pass into a constructor of the object.

To break such circular dependency it is better to create an object
without attributes, scan attributes as planned, and only then add
the attributes to the object.