Port a list of show years from fixed to generated
Currently, a per year list of shows is generated from external paths and
captions enumerated in res/show_years.yml. This commit replaces it with
automatic scanning for a "t=" parameter in photo names to take a photo
as a title card for a year between
sections.shows.get_root_artifact_list().FIRST_SHOW_YEAR and a year when
a script is started.