Change dog list photo dimensions
Currently all photos in dogs/, females/, and males/ lists are resized
to fit into 200px width because title photos are supposed to be
horizontal with dogs in a stand.

However, a temporary vertical photo 2015/09052 with sitting Yunaru broke
the layout because its preview height is three times larger than others.

So this commit changes dog card preview sizing to 152px height as for
the rest of a site. Also, it reduces file count of deployed photos by
getting rid of *-w558.jpg versions.