Remove explicit notion of .jpg files as binary
According to [an answer of Chris Johnsen to "How to determine if Git
handles a file as binary or as text?" (licensed under CC BY-SA
4)](https://stackoverflow.com/a/6134127/2553424):

> builtin_diff() calls diff_filespec_is_binary() which calls
> buffer_is_binary() which checks for any occurrence of a zero byte
> (NUL "character") in the first 8000 bytes (or the entire length if
> shorter).

So we do not need to explicitly specify JFIF and PNG images as binary
via .gitignore. JFIF always has a zero byte by offset 0x4 as a part of
a two byte header length (00 10). PNG has a zero byte as a part of IHDR
fixed length (00 00 00 0D).