Clarify non-encapsulated nature of Photo fields
Currently a scanning through source code may give an illusion that
Photo._image and Photo._dogs are written once, in a constructor and can
not be modified again due to encapsulation. However, that is not true,
so this commit moves all initialization from the constructor. Though
such leakage happens only while construction and only from a method that
created that object, we need to explicitly specify that a field is
created only by side modifications because some methods always return
None otherwise.