Log generation not of documents, but of photos
The longest operation a logging was introduced for in the first place is
photo generation. Before introduction of promise-based generation it was
performed in page generators, so their invocation was logged for having
a context of exceptions. Currently, however, pages are generated swiftly
being supplied by photos in a separate pass.

So this commit replaces logging of per-page generator invocation
to logging of promise keeping (fulfillment) for each source photo file.